+++ src/PAOFLOW/gen/paoflow_driver.py
@@ -19,7 +19,7 @@ Two further workflows generate multi-phase driver scripts:
 
 3. **Harmonic phonons** (``main.phonon.py``) -- finite-displacement phonons via
    phonopy (generate supercells -> run pw.x -> analyse forces).
-4. **Electron-phonon** (``main.elphon.py``) -- the atomic-orbital (Agapito &
+4. **Electron-phonon** (``main.elphon.py``) -- the pseudo-atomic-orbital (Agapito &
    Bernardi) interpolation of QE's DFPT coupling into ``alpha^2F`` / ``lambda`` /
    ``Tc`` (write the ph.x phonon + AHC inputs -> run QE -> analyse).
 
@@ -1247,12 +1247,12 @@ def build_phonon_script(cfg):
 
 
 # --------------------------------------------------------------------------- #
-# Electron-phonon (AO route) workflow builder
+# Electron-phonon (PAO route) workflow builder
 # --------------------------------------------------------------------------- #
 ELPHON_TEMPLATE = r'''#!/usr/bin/env python3
-"""PAOFLOW electron-phonon (Eliashberg) workflow -- AO route (generated by paoflow_gen.py).
+"""PAOFLOW electron-phonon (Eliashberg) workflow -- PAO route (generated by paoflow_gen.py).
 
-Atomic-orbital (Agapito & Bernardi, Phys. Rev. B 97, 235146 (2018)) interpolation
+Pseudo-atomic-orbital (Agapito & Bernardi, Phys. Rev. B 97, 235146 (2018)) interpolation
 of Quantum ESPRESSO's DFPT electron-phonon coupling.  PAOFLOW reads QE's *full*
 coarse-grid coupling (no potential reconstruction), rotates it into the PAO gauge
 and Wigner-Seitz interpolates electrons + vertex to a dense grid for alpha^2F,
@@ -1261,7 +1261,7 @@ lambda, omega_log and Tc.
 Two phases:
 
     python main.elphon.py inputs    # write the ph.x phonon + AHC input templates
-    python main.elphon.py analyse   # AO interpolation -> alpha^2F, lambda, Tc
+    python main.elphon.py analyse   # PAO interpolation -> alpha^2F, lambda, Tc
 
 Between them, run the two QE ph.x steps in the same outdir (typically on HPC):
 
@@ -1280,7 +1280,7 @@ import sys
 import numpy as np
 
 from PAOFLOW import PAOFLOW
-from PAOFLOW.elphon.do_ao_eph import eliashberg_from_qe_coupling
+from PAOFLOW.elphon.do_pao_eph import eliashberg_from_qe_coupling
 from PAOFLOW.elphon.elph_bloch import read_nscf
 
 
@@ -1362,7 +1362,7 @@ def inputs():
 
 
 def analyse():
-    """Phase 2: AO interpolation of the QE coupling -> alpha^2F, lambda, Tc."""
+    """Phase 2: PAO interpolation of the QE coupling -> alpha^2F, lambda, Tc."""
     if not os.path.isdir(SAVEDIR):
         sys.exit('%s not found. Run pw.x (nscf) first.' % SAVEDIR)
     if not os.path.isdir(COUPLING_DIR):
@@ -1393,7 +1393,7 @@ def analyse():
     )
 
     kB = 8.617333262e-5  # eV/K
-    print('AO-route Eliashberg (%s, source=%s, Nk=%d):' % (PREFIX, SOURCE, NK_DENSE))
+    print('PAO-route Eliashberg (%s, source=%s, Nk=%d):' % (PREFIX, SOURCE, NK_DENSE))
     print('  N(E_F)   = %.3f states/spin/Ry' % out['dos_ef'].mean())
     print('  lambda   = %.4f' % out['lambda'])
     print('  <w_log>  = %.1f K' % (out['omega_log'] / kB))
@@ -1408,7 +1408,7 @@ def analyse():
 
 
 def main():
-    parser = argparse.ArgumentParser(description='PAOFLOW electron-phonon (AO) workflow.')
+    parser = argparse.ArgumentParser(description='PAOFLOW electron-phonon (PAO) workflow.')
     parser.add_argument('phase', nargs='?', choices=['inputs', 'analyse', 'all'],
                         default='all', help='Workflow phase to run (default: all).')
     args = parser.parse_args()
@@ -1424,7 +1424,7 @@ if __name__ == '__main__':
 
 
 def build_elphon_script(cfg):
-    """Assemble a main.elphon.py electron-phonon (AO route) workflow from the config."""
+    """Assemble a main.elphon.py electron-phonon (PAO route) workflow from the config."""
     subs = {
         '__SAVEDIR__': repr(cfg['savedir']),
         '__OUTPUTDIR__': repr(cfg['outputdir']),
@@ -3114,7 +3114,7 @@ def collect_phonon(common):
 
 
 def collect_elphon(common):
-    """Prompt for the electron-phonon (AO route) configuration."""
+    """Prompt for the electron-phonon (PAO route) configuration."""
     cfg = dict(common)
     cfg['source'] = ask_choice(
         'Coupling source (ahc = unpatched QE, norm-conserving; '
